+++ test/cases/wodule.js
@@ -41,7 +41,8 @@ describe('#wodule', function() {
         });
 
         it('module.start()', function() {
-            return wodule.start().tap(function() {
+            return wodule.start().then(function(result) {
+                result.should.be.true();
                 wodule.initialized.should.be.true();
                 wodule.running.should.be.true();
                 Wodule.prototype._start.should.be.calledOnce();
@@ -49,7 +50,8 @@ describe('#wodule', function() {
         });
 
         it('module.stop()', function() {
-            return wodule.stop().tap(function() {
+            return wodule.stop().then(function(result) {
+                result.should.be.true();
                 wodule.initialized.should.be.true();
                 wodule.running.should.be.false();
                 Wodule.prototype._stop.should.be.calledOnce();
@@ -59,7 +61,7 @@ describe('#wodule', function() {
         it('module.exit()', function() {
             wodule.exit(3);
             wodule.exit(2);
-            var p = wodule.exit(1).tap(function() {
+            return wodule.exit(1).then(function() {
                 wodule.initialized.should.be.true();
                 wodule.running.should.be.false();
 
@@ -68,7 +70,6 @@ describe('#wodule', function() {
                 process.exit.should.be.calledOnce();
                 process.exit.should.be.calledWith(3);
             });
-            return p;
         });
     });
 
@@ -94,6 +95,95 @@ describe('#wodule', function() {
                 });
             }, function(err) {
                 return err.message === 'The method must be a Function!';
+            });
+        });
+    });
+
+    describe('test functions with wodule', function() {
+        var wodule;
+        before(function() {
+            wodule = new Wodule();
+        });
+
+        describe('init', function() {
+            after(function() {
+                delete wodule._init;
+            });
+
+            it('_init throws an error', function() {
+                var message = 'exception occurred in _init';
+                wodule._init = function() {
+                    throw new Error(message);
+                };
+
+                should.throws(function(err) {
+                    wodule.init();
+                }, function(err) {
+                    return err.message === message;
+                });
+                wodule.initialized.should.be.false();
+            });
+
+            it('_init return false', function() {
+                wodule._init = function() {
+                    return false;
+                };
+
+                should.throws(function(err) {
+                    wodule.init();
+                }, function(err) {
+                    return err.message === 'module._init failed';
+                });
+                wodule.initialized.should.be.false();
+            });
+
+            it('_init return undefined', function() {
+                wodule._init = function() {};
+                wodule.init();
+                wodule.initialized.should.be.true();
+            });
+        });
+
+        describe('start', function() {
+            after(function() {
+                delete wodule._start;
+            });
+
+            it('_start throws an error', function() {
+                var message = 'exception occurred in _start';
+                wodule._start = function() {
+                    throw new Error(message);
+                };
+
+                return wodule.start().catch(function(err) {
+                    wodule.running.should.be.false();
+                    err.message.should.equal(message);
+                });
+            });
+
+            it('_start callback(err)', function() {
+                var message = 'exception occurred in _start';
+                wodule._start = function(callback) {
+                    callback(new Error(message));
+                };
+
+                return wodule.start().catch(function(err) {
+                    wodule.running.should.be.false();
+                    err.message.should.equal(message);
+                });
+            });
+
+            it('_start callback()', function() {
+                wodule._start = function(callback) {
+                    callback();
+                };
+
+                return wodule.start()
+                    .then(function(result) {
+                        wodule.initialized.should.be.true();
+                        wodule.running.should.be.true();
+                        result.should.be.true();
+                    });
             });
         });
     });