+++ ChangeLog.md
@@ -1,3 +1,17 @@
+<a name="0.1.0"></a>
+# 0.1.0 (2016-01-01)
+
+
+### Bug Fixes
+
+* **util:** remote timeout for returnOrCallback promise ([219ef98](https://github.com/adoyle-h/wodule/commit/219ef98))
+
+### Features
+
+* init synchronously and change start\stop detail ([1571111](https://github.com/adoyle-h/wodule/commit/1571111))
+
+
+
 <a name="0.0.3"></a>
 ## 0.0.3 (2015-12-29)